Refactor XHCI interrupt handling to use task notification
This update changes the way that XHCI interrupts are handled. Instead of being queued directly, a message is created and sent to a newly created 'usb_handler' task. A function to get a task_id by the name has also been added, allowing the XHCI interrupt to be processed by the relevant task. The handling of the XHCI interrupt has been moved to a dedicated USB handling task, improving modularization and isolation of the system's components.