Define INTERRUPT_TASK_ID and refine message validation in ipc
The constant INTERRUPT_TASK_ID was added to types.hpp for identifying interrupt tasks. Further, the validation in the method 'send_message' in 'ipc.cpp' was enhanced to not only reject invalid destination ids, but also prevent tasks from sending messages to themselves. In 'handlers.cpp', the interrupt handler is now using the defined INTERRUPT_TASK_ID for their messages.