Implement SYS_EXIT syscall and adjust task kernel stack
The SYS_EXIT system call was added to support clean termination of tasks. Also, the format of `call_userland` was modified to save callee-saved registers and the kernel stack top of the current task.  Further, the 'syscall_number' was saved before 'handle_syscall' and was used to handle task termination directly in the syscall entry code.