Implement kill_userland function and improve ELF handling
Added a function to handle userland exit in response to specific interrupts, and improved handling of non-ELF files in the FAT module and terminal output. Also, simplified the use of templates in terminal.hpp and terminal.cpp for printk function. Additionally, the vector class was included in page.cpp to manage memory pages.